shared/mqtt/paho_basic_callbacks/pyproject.toml
[tool.poetry]
name = "paho-basic-callbacks"
version = "0.1.0"
description = ""
authors = ["Shahar Tal <chapo12345@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
